PreprocessingPipelineJuML

Documentation for PreprocessingPipelineJuML.

Preprocessing_Pipeline_JuML.expand_contractions

Preprocessing_Pipeline_JuML.get_vocab_dict

Preprocessing_Pipeline_JuML.mask_numbers

Preprocessing_Pipeline_JuML.one_hot_encoding

Preprocessing_Pipeline_JuML.remove_noise

Preprocessing_Pipeline_JuML.remove_stop_words

Preprocessing_Pipeline_JuML.standardize_text

Preprocessing_Pipeline_JuML.expand_contractions — Method

expand_contractions(input::NlpPipe) -> NlpPipe

Expand contractions in the input text. This function expands common English contractions.

Arguments

input::NlpPipe: A NlpPipe object containing the corpus to expand contractions in.

Returns

output::NlpPipe: A new NlpPipe object with the contractions expanded in the corpus.

Preprocessing_Pipeline_JuML.get_vocab_dict — Method

get_vocab_dict(vocab)

Create a Dict with idx in vector as values and words as keys

Preprocessing_Pipeline_JuML.mask_numbers — Method

number_to_word(input)

Allows user to preprocess numbers within a string. Each number, independent of length gets replaced with word: *number*, 
inputs: Strings and Vector{String}
output: Strings and Vector{String} respectively

Preprocessing_Pipeline_JuML.one_hot_encoding — Method

one_hot_encoding(x)

Create a one-hot-encoding out of given TokenizedNlpPipe

#Examples:
```julia-repl
julia> pipe = TokenizedNlpPipe([["I", "love", "Julia"], ["Julia", "is", "awesome"]])
julia> one_hot(pipe)
VectorizedNlpPipe([[[1 0 0 0 0; 0 1 0 0 0; 0 0 1 0 0], [0 0 1 0 0; 0 0 0 1 0; 0 0 0 0 1]]], Dict("love" => 2, "Julia" => 3, "I" => 1, "is" => 4, "awesome" => 5))
```

Preprocessing_Pipeline_JuML.remove_noise — Method

remove_noise(pipe::NlpPipe)

Removes noise from the corpus. Noise includes HTML tags, URLs, email addresses, file paths, special characters, dates & times.

Arguments

pipe::NlpPipe: The NlpPipe object with a corpus to remove noise from

Returns

A new NlpPipe object with the noise removed from the corpus

Preprocessing_Pipeline_JuML.remove_stop_words — Method

remove_stop_words(pipe::TokenizedNlpPipe; language::String="en", stop_words::Set{String}=Set{String}())

Removes predefined stopwords. You can access the stop words for a given language using the language name or ISO 639 code. For example, to get the stop words for English, you can use stopwords["eng"], stopwords["en"], or stopwords["English"].

Stop words sourced from https://github.com/guo-yong-zhi/StopWords.jl/blob/main/README.md.

Parameters

pipe: TokenizedNlpPipe

language: String = "en"

stop_words: Set{String} = Set{String}()

Returns

A new TokenizedNlpPipestruct with the stop words removed from the tokens.

Examples

NlpPipe(["This is a test"]) |> tokenize |> remove_stop_words |> pipe -> pipe.tokens # [["This", "test"]]
NlpPipe(["This is a test"]) |> tokenize |> remove_stop_words(stop_words=["test"]) |> pipe -> pipe.tokens # [["This", "is", "a"]]

Preprocessing_Pipeline_JuML.standardize_text — Method

standardize_text(input::NlpPipe)

Standardizes the text by converting it to lowercase and replacing unusual characters with their standard counterparts.